requirements.txt
Django==4.1.5
mysqlclient==2.1.1
djangorestframework==3.14.0
drf-yasg==1.21.4
paramiko==3.0.0
django-bootstrap5==22.2